Update ProcessDocumentation.md
Made a few small tweaks. 
Also wonder if we should use the "What is Exploration" explanation from the reports to keep it consistent?
Could we test this with someone who knows nothing about exploration at CDS to see if it makes sense to someone who hasn't lived in our world for the last few months?